refactor: extract time formatting logic into a utility function
This commit introduces a new utility function `formatTimeLeft` in `time-formatting.ts` to handle the countdown timer logic for discount expiration. The `timeLeftText` getters in both `ReferralDiscountNoticeComponent` and `SignupDiscountNoticeComponent` have been updated to utilize this new function, simplifying the code and improving maintainability.
