+++ app/components/pay-page/signup-discount-notice.ts
@@ -3,7 +3,7 @@ import PromotionalDiscountModel from 'codecrafters-frontend/models/promotional-d
 import iconImage from '/assets/images/icons/money.svg';
 import { service } from '@ember/service';
 import type TimeService from 'codecrafters-frontend/services/time';
-
+import { formatTimeLeft } from 'codecrafters-frontend/utils/time-formatting';
 interface Signature {
   Element: HTMLDivElement;
 
@@ -18,24 +18,7 @@ export default class SignupDiscountNoticeComponent extends Component<Signature> 
   @service declare time: TimeService;
 
   get timeLeftText() {
-    const distanceInSeconds = Math.floor((this.args.discount.expiresAt.getTime() - this.time.currentTime.getTime()) / 1000);
-    const hoursLeft = Math.floor(distanceInSeconds / 60 / 60);
-    const minutesLeft = Math.floor(distanceInSeconds / 60) - hoursLeft * 60;
-    const secondsLeft = distanceInSeconds - hoursLeft * 60 * 60 - minutesLeft * 60;
-
-    const hoursLeftStr = `${hoursLeft.toString().padStart(2, '0')}h`;
-    const minutesLeftStr = `${minutesLeft.toString().padStart(2, '0')}m`;
-    const secondsLeftStr = `${secondsLeft.toString().padStart(2, '0')}s`;
-
-    if (hoursLeft > 0) {
-      return `${hoursLeftStr}:${minutesLeftStr}:${secondsLeftStr}`;
-    }
-
-    if (minutesLeft > 0) {
-      return `${minutesLeftStr}:${secondsLeftStr}`;
-    }
-
-    return `${secondsLeftStr}`;
+    return formatTimeLeft(this.args.discount.expiresAt, this.time.currentTime);
   }
 }
